1) Get DOI from doi/query database which is created by query procedure 2) jEvalMacros updated to reuse the data from edited authors workbook 3) proc_doi (query procedure): still testing/fixing

+++ proc/proc_doi/query/shs/readconfig.sh
@@ -31,9 +31,14 @@ then
     echo nothing> $APP_STFILE
 fi
 
+if [ ! -f $CIPFILE ]
+then
+    echo GENERATING $CIPFILE
+    echo ARTIGO.*=$ARTIGO.*     >  $CIPFILE
+    echo ISSUE.*=$ISSUEDB.*     >> $CIPFILE
+    echo TITLE.*=$TITLEDB.*     >> $CIPFILE
+    echo QUERY.*=$QUERYDB.*     >> $CIPFILE
+    echo QUERYLOG.*=$QUERYLOGDB.*     >> $CIPFILE
+    echo GENERATED $CIPFILE
 
-echo ARTIGO.*=$ARTIGO.*     >  $CIPFILE
-        echo ISSUE.*=$ISSUEDB.*     >> $CIPFILE
-        echo TITLE.*=$TITLEDB.*     >> $CIPFILE
-        echo QUERY.*=$QUERYDB.*     >> $CIPFILE
-        echo QUERYLOG.*=$QUERYLOGDB.*     >> $CIPFILE
+fi
